fix dedicated server not launching
